user_service/requirements.txt
Django==4.2.22
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
